Function Update

+++ QuanLyInternet/FormHoSoKhachHang.cs
@@ -37,29 +37,47 @@ namespace QuanLyInternet
 
         private void ambiance_Button_13_Click(object sender, EventArgs e)
         {
-            int maHDindex = dataGridView1.Columns["MaHopDong"].Index;
-            //dataGridView1.SelectedRows;
-            int rowIndex = dataGridView1.CurrentCell.RowIndex;
-            string maHD = dataGridView1[maHDindex, rowIndex].Value.ToString();
-            FormHienThiKhachHang hienThiKhachHang = new FormHienThiKhachHang(maHD);
-            hienThiKhachHang.Show();
-        }
-
-        private void ambiance_Button_12_Click(object sender, EventArgs e)
-        {
-            int maHDindex = dataGridView1.Columns["MaHopDong"].Index;
-            //dataGridView1.SelectedRows;
-            int rowIndex = dataGridView1.CurrentCell.RowIndex;
-            string maHD = dataGridView1[maHDindex, rowIndex].Value.ToString();
             try
             {
-                Database.GetInstance.HopDong.deleteHopDong(maHD);
-                reload();
+                int maHDindex = dataGridView1.Columns["MaHopDong"].Index;
+                //dataGridView1.SelectedRows;
+                int rowIndex = dataGridView1.CurrentCell.RowIndex;
+                string maHD = dataGridView1[maHDindex, rowIndex].Value.ToString();
+                FormHienThiKhachHang hienThiKhachHang = new FormHienThiKhachHang(maHD);
+                hienThiKhachHang.Show();
             }
             catch
             {
  
             }
+            
+        }
+
+        private void ambiance_Button_12_Click(object sender, EventArgs e)
+        {
+
+            DialogResult dialogResult = MessageBox.Show("Bạn có chắc chắn muốn xóa?", "Xác Nhận", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                int maHDindex = dataGridView1.Columns["MaHopDong"].Index;
+                //dataGridView1.SelectedRows;
+                int rowIndex = dataGridView1.CurrentCell.RowIndex;
+                string maHD = dataGridView1[maHDindex, rowIndex].Value.ToString();
+                try
+                {
+                    Database.GetInstance.HopDong.deleteHopDong(maHD);
+                    reload();
+                }
+                catch
+                {
+
+                }
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+
+            }
+            
             
         }
